requirements.txt
Flask
Gunicorn
mysql-connector-python
python-dotenv
matplotlib
networkx